Module_7/requirements.txt
openai==1.42.0       
faiss-cpu==1.8.0     
PyPDF2==3.0.1        
numpy==1.26.4       
python-dotenv==1.0.1 